added more user feature

+++ application/controllers/orders.php
@@ -9,10 +9,26 @@ class Orders extends CI_Controller
         // $this->output->enable_profiler();
     }
 
+    public function orders_route(){
+        if ($this->session->userdata('admin_level')!=9){
+            redirect('/');
+        }else {
+            $laka['orders'] = $this->order->get_all_orders();
+            $this->load->view('orders_view', $laka);
+        }
+    }
+
     public function add_order(){
+        if($this->session->userdata('userID')){
+            $userAddress = $this->usermodel->get_user($this->session->userdata('userID'))['address'];
+        }
         if (!$this->session->userdata('userID')){
             redirect('/login');
-        } else {
+        } else if (empty($this->session->userdata('cart'))){
+            redirect('/cart');
+        }else if ($userAddress == NULL) {
+            redirect('/users/dashboard/'.$this->session->userdata('userID'));
+        }else {
             $order = [
                 "user_id" => $this->session->userdata('userID'),
                 "status" => 'Processing'
@@ -25,9 +41,14 @@ class Orders extends CI_Controller
     }
 
     public function show_order($orderID){
-        $laka['order'] = $this->order->get_order($orderID);
-        $laka['orderID'] = $orderID;
-        $this->load->view('showorder', $laka);
+        if ($this->session->userdata('admin_level')==9){
+            $laka['order'] = $this->order->get_order($orderID);
+            $laka['orderID'] = $orderID;
+            $this->load->view('showorder', $laka);
+        } else {
+            redirect('/');
+        }
+
     }
 
     public function edit_order($orderID){
@@ -36,8 +57,13 @@ class Orders extends CI_Controller
     }
 
     public function delete_order($orderID){
-        $this->order->delete_order($orderID);
-        redirect('/admin/orders');
+        if ($this->session->userdata('admin_level')==9){
+            $this->order->delete_order($orderID);
+            redirect('/admin/orders');
+        } else {
+            redirect('/');
+        }
+
     }
 
 }
